chibi_core::input

Enum Command

pub enum Command {

    SendPrompt {
        prompt: String,
    },
    ListContexts,
    ListCurrentContext,
    DestroyContext {
        name: Option<String>,
    },
    ArchiveHistory {
        name: Option<String>,
    },
    CompactContext {
        name: Option<String>,
    },
    RenameContext {
        old: Option<String>,
        new: String,
    },
    ShowLog {
        context: Option<String>,
        count: isize,
    },
    Inspect {
        context: Option<String>,
        thing: Inspectable,
    },
    SetSystemPrompt {
        context: Option<String>,
        prompt: String,
    },
    SetModel {
        context: Option<String>,
        model: String,
    },
    RunPlugin {
        name: String,
        args: Vec<String>,
    },
    CallTool {
        name: String,
        args: Vec<String>,
    },
    ClearCache {
        name: Option<String>,
    },
    CleanupCache,
    CheckInbox {
        context: String,
    },
    CheckAllInboxes,
    ModelMetadata {
        model: String,
        full: bool,
    },
    ShowHelp,
    ShowVersion,
    NoOp,

}

What operation to perform (mutually exclusive commands)

Variants

SendPrompt

Send a prompt to the LLM

Fields

prompt: String

ListContexts

List all contexts (-L)

ListCurrentContext

Show current context info (-l)

DestroyContext

Destroy a context (-d/-D)

Fields

name: Option<String>

ArchiveHistory

Archive a contextโs history (-a/-A)

Fields

name: Option<String>

CompactContext

Compact a context (-z/-Z)

Fields

name: Option<String>

RenameContext

Rename a context (-r/-R)

Fields

old: Option<String>

new: String

ShowLog

Show log entries (-g/-G)

Fields

context: Option<String>

count: isize

Inspect

Inspect something (-n/-N)

Fields

context: Option<String>

thing: Inspectable

SetSystemPrompt

Set system prompt (-y/-Y)

Fields

context: Option<String>

prompt: String

SetModel

Set model for a context (-m/โset-model, -M/โset-model-for-context)

Fields

context: Option<String>

model: String

RunPlugin

Run a plugin directly (-p)

Fields

name: String

args: Vec<String>

CallTool

Call a tool directly (-P)

Fields

name: String

args: Vec<String>

ClearCache

Clear tool cache (โclear-cache/โclear-cache-for)

Fields

name: Option<String>

CleanupCache

Cleanup old cache entries (โcleanup-cache)

CheckInbox

Check inbox for a specific context and process any messages (-B)

Fields

context: String

CheckAllInboxes

Check all context inboxes and process any messages (-b)

ModelMetadata

Show model metadata from registry (-m/-M)

Fields

model: String

full: bool

ShowHelp

Show help

ShowVersion

Show version

NoOp

No operation - context switch only, no action